SHK-4545: Exposed custom feedback types

+++ android/src/main/java/com/shakebugs/react/utils/Mapper.java
@@ -1,16 +1,21 @@
 package com.shakebugs.react.utils;
 
+import android.content.Context;
+
 import com.facebook.react.bridge.ReadableArray;
 import com.facebook.react.bridge.ReadableMap;
 import com.facebook.react.bridge.ReadableMapKeySetIterator;
 import com.facebook.react.bridge.ReadableType;
+import com.facebook.react.bridge.WritableArray;
 import com.facebook.react.bridge.WritableMap;
+import com.facebook.react.bridge.WritableNativeArray;
 import com.facebook.react.bridge.WritableNativeMap;
 import com.shakebugs.shake.LogLevel;
 import com.shakebugs.shake.ShakeReportConfiguration;
 import com.shakebugs.shake.ShakeScreen;
 import com.shakebugs.shake.internal.data.NetworkRequest;
 import com.shakebugs.shake.internal.data.NotificationEvent;
+import com.shakebugs.shake.report.FeedbackType;
 import com.shakebugs.shake.report.ShakeFile;
 
 import java.util.ArrayList;
@@ -19,7 +24,15 @@ import java.util.List;
 import java.util.Map;
 
 public class Mapper {
-    public static LogLevel mapToLogLevel(ReadableMap logLevelMap) {
+
+    private final Context context;
+
+    public Mapper(Context context) {
+        this.context = context;
+    }
+
+    public LogLevel mapToLogLevel(ReadableMap logLevelMap) {
+
         LogLevel logLevel = LogLevel.INFO;
 
         try {
@@ -31,7 +44,8 @@ public class Mapper {
         return logLevel;
     }
 
-    public static ShakeScreen mapToShakeScreen(ReadableMap shakeScreenMap) {
+    public ShakeScreen mapToShakeScreen(ReadableMap shakeScreenMap) {
+
         ShakeScreen shakeScreen = ShakeScreen.HOME;
 
         try {
@@ -43,7 +57,8 @@ public class Mapper {
         return shakeScreen;
     }
 
-    public static ShakeReportConfiguration mapToConfiguration(ReadableMap configurationMap) {
+    public ShakeReportConfiguration mapToConfiguration(ReadableMap configurationMap) {
+        if (configurationMap == null) return null;
 
         boolean blackBoxData = configurationMap.getBoolean("blackBoxData");
         boolean activityHistoryData = configurationMap.getBoolean("activityHistoryData");
@@ -59,22 +74,76 @@ public class Mapper {
         return configuration;
     }
 
-    public static List<ShakeFile> mapToShakeFiles(ReadableArray filePaths) {
+    public List<ShakeFile> mapArrayToShakeFiles(ReadableArray filePaths) {
+        if (filePaths == null) return null;
+
         List<ShakeFile> shakeFiles = new ArrayList<>();
         for (int i = 0; i < filePaths.size(); i++) {
             ReadableMap fileMap = filePaths.getMap(i);
 
             String filePath = fileMap.getString("path");
             String fileName = fileMap.getString("name");
+            fileName = Files.removeExtension(fileName);
 
-            String name = Files.removeExtension(fileName);
-
-            shakeFiles.add(new ShakeFile(name, filePath));
+            if (fileName != null && filePath != null) {
+                shakeFiles.add(new ShakeFile(fileName, filePath));
+            }
         }
         return shakeFiles;
     }
 
-    public static NetworkRequest mapToNetworkRequest(ReadableMap object) {
+    public List<FeedbackType> mapArrayToFeedbackTypes(ReadableArray feedbackTypesArray) {
+        if (feedbackTypesArray == null) return null;
+
+        List<FeedbackType> feedbackTypes = new ArrayList<>();
+        for (int i = 0; i < feedbackTypesArray.size(); i++) {
+            ReadableMap feedbackTypeMap = feedbackTypesArray.getMap(i);
+
+            String title = feedbackTypeMap.getString("title");
+            String tag = feedbackTypeMap.getString("tag");
+            String icon = feedbackTypeMap.getString("icon");
+
+            int resourceId = context.getResources().getIdentifier(
+                    icon, "drawable", context.getPackageName());
+
+            if (title != null && tag != null) {
+                feedbackTypes.add(new FeedbackType(resourceId, title, tag));
+            }
+        }
+        return feedbackTypes;
+    }
+
+    public WritableArray mapFeedbackTypesToArray(List<FeedbackType> feedbackTypes) {
+        if (feedbackTypes == null) return null;
+
+        WritableArray feedbackTypesArray = new WritableNativeArray();
+
+        for (FeedbackType feedbackType : feedbackTypes) {
+
+            if (feedbackType.getTitle() == null) {
+                feedbackType.setTitle(context.getString(feedbackType.getTitleRes()));
+            }
+
+            String feedbackIcon = "";
+            try {
+                feedbackIcon = context.getResources().getResourceEntryName(feedbackType.getIcon());
+            } catch (Exception ignore) {
+            }
+
+            WritableMap feedbackTypeMap = new WritableNativeMap();
+            feedbackTypeMap.putString("title", feedbackType.getTitle());
+            feedbackTypeMap.putString("tag", feedbackType.getTag());
+            feedbackTypeMap.putString("icon", feedbackIcon);
+
+            feedbackTypesArray.pushMap(feedbackTypeMap);
+        }
+
+        return feedbackTypesArray;
+    }
+
+    public NetworkRequest mapToNetworkRequest(ReadableMap object) {
+        if (object == null) return null;
+
         NetworkRequest networkRequest = new NetworkRequest();
         networkRequest.setUrl(object.getString("url"));
         networkRequest.setMethod(object.getString("method"));
@@ -89,7 +158,9 @@ public class Mapper {
         return networkRequest;
     }
 
-    public static NotificationEvent mapToNotificationEvent(ReadableMap object) {
+    public NotificationEvent mapToNotificationEvent(ReadableMap object) {
+        if (object == null) return null;
+
         String id = object.hasKey("id") && !object.isNull("id") ? object.getString("id") : "";
         String title = object.hasKey("title") && !object.isNull("title") ? object.getString("title") : "";
         String description = object.hasKey("description") && !object.isNull("description") ? object.getString("description") : "";
@@ -102,7 +173,8 @@ public class Mapper {
         return notificationEvent;
     }
 
-    public static WritableMap notificationEventToMap(NotificationEvent notificationEvent) {
+    public WritableMap notificationEventToMap(NotificationEvent notificationEvent) {
+
         int id = notificationEvent.getId();
         String title = notificationEvent.getTitle() == null ? "" : notificationEvent.getTitle();
         String description = notificationEvent.getDescription() == null ? "" : notificationEvent.getDescription();
@@ -115,7 +187,9 @@ public class Mapper {
         return map;
     }
 
-    public static Map<String, String> mapToUserMetadata(ReadableMap metadata) {
+    public Map<String, String> mapToUserMetadata(ReadableMap metadata) {
+        if (metadata == null) return null;
+
         Map<String, Object> map = toMap(metadata);
 
         Map<String, String> stringMap = new HashMap<>();
@@ -130,7 +204,8 @@ public class Mapper {
         return stringMap;
     }
 
-    private static Map<String, Object> toMap(ReadableMap readableMap) {
+    private Map<String, Object> toMap(ReadableMap readableMap) {
+
         Map<String, Object> map = new HashMap<>();
         ReadableMapKeySetIterator iterator = readableMap.keySetIterator();
 
@@ -163,7 +238,7 @@ public class Mapper {
         return map;
     }
 
-    private static Map<String, String> toStringMap(ReadableMap readableMap) {
+    private Map<String, String> toStringMap(ReadableMap readableMap) {
         Map<String, Object> map = toMap(readableMap);
 
         Map<String, String> stringMap = new HashMap<>();
@@ -176,7 +251,7 @@ public class Mapper {
         return stringMap;
     }
 
-    private static Object[] toArray(ReadableArray readableArray) {
+    private Object[] toArray(ReadableArray readableArray) {
         Object[] array = new Object[readableArray.size()];
 
         for (int i = 0; i < readableArray.size(); i++) {
